重构:main_window和info_window单独model
